requirements.txt
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0

pyspark==3.4.1
requests==2.31.0
urllib3==1.26.16